Add observations to plotting
